Update part2c.md
this might be nit picking but the phrase ‘so-called’ is used to my mind incorrectly and frequently throughout the course.  the meaning of ‘so-called’, as I understand it, makes the sentences that use it a bit misleading.

saying ‘so-called promises’ expresses doubt that promises are correctly named.

to me it reads:
people call them promises but they really shouldn’t.

the most positive interpretation might me that they are called that by convention but it’s not their real name and they are perhaps not called that by everyone.

when in fact they are promises and that is their real name.